requirements.txt
# Flask и расширения
Flask==3.0.0
Flask-CORS==4.0.0

# Работа с переменными окружения
python-dotenv==1.0.0

# HTTP клиент для запросов к Green API
requests==2.31.0

# Тестирование
pytest==7.4.3
pytest-flask==1.3.0

# WSGI сервер для продакшена (опционально)
gunicorn==21.2.0
